Avoid unnecessary call to `touch`
This isn't a refactoring, because there aren't any before/after tests.
However, we're returning from the function early, if the two arguments
point to the same file. This avoids an unnecessary call to `touch`.

+++ sqlrunner/main.py
@@ -94,8 +94,8 @@ def write_results(results, f_path):
 
 @write_results.register
 def _(results: pathlib.Path, f_path):
+    if f_path.exists() and f_path.samefile(results):
+        return
+
     touch(f_path)
-    try:
-        shutil.copy(results, f_path)
-    except shutil.SameFileError:
-        pass
+    shutil.copy(results, f_path)
